+++ src/Twilio/Rest/Messaging/V1/ServiceContext.php
@@ -27,6 +27,7 @@ use Twilio\Serialize;
 use Twilio\Rest\Messaging\V1\Service\AlphaSenderList;
 use Twilio\Rest\Messaging\V1\Service\PhoneNumberList;
 use Twilio\Rest\Messaging\V1\Service\UsAppToPersonUsecaseList;
+use Twilio\Rest\Messaging\V1\Service\ChannelSenderList;
 use Twilio\Rest\Messaging\V1\Service\ShortCodeList;
 use Twilio\Rest\Messaging\V1\Service\UsAppToPersonList;
 
@@ -35,18 +36,21 @@ use Twilio\Rest\Messaging\V1\Service\UsAppToPersonList;
  * @property AlphaSenderList $alphaSenders
  * @property PhoneNumberList $phoneNumbers
  * @property UsAppToPersonUsecaseList $usAppToPersonUsecases
+ * @property ChannelSenderList $channelSenders
  * @property ShortCodeList $shortCodes
  * @property UsAppToPersonList $usAppToPerson
  * @method \Twilio\Rest\Messaging\V1\Service\ShortCodeContext shortCodes(string $sid)
  * @method \Twilio\Rest\Messaging\V1\Service\UsAppToPersonContext usAppToPerson(string $sid)
  * @method \Twilio\Rest\Messaging\V1\Service\PhoneNumberContext phoneNumbers(string $sid)
  * @method \Twilio\Rest\Messaging\V1\Service\AlphaSenderContext alphaSenders(string $sid)
+ * @method \Twilio\Rest\Messaging\V1\Service\ChannelSenderContext channelSenders(string $sid)
  */
 class ServiceContext extends InstanceContext
     {
     protected $_alphaSenders;
     protected $_phoneNumbers;
     protected $_usAppToPersonUsecases;
+    protected $_channelSenders;
     protected $_shortCodes;
     protected $_usAppToPerson;
 
@@ -204,6 +208,21 @@ class ServiceContext extends InstanceContext
         }
 
         return $this->_usAppToPersonUsecases;
+    }
+
+    /**
+     * Access the channelSenders
+     */
+    protected function getChannelSenders(): ChannelSenderList
+    {
+        if (!$this->_channelSenders) {
+            $this->_channelSenders = new ChannelSenderList(
+                $this->version,
+                $this->solution['sid']
+            );
+        }
+
+        return $this->_channelSenders;
     }
 
     /**